pyproject.toml
[tool.poetry]
name = "fizzbotz"
version = "0.1.0"
description = "A Discord bot by Fission#1337"
authors = ["Fission"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
aiofiles = "^0.4.0"
aiohttp = "^3.4"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", branch = "rewrite"}
pynacl = "^1.2"
python-dotenv = "^0.9.1"
uvloop = {version = "^0.11.2", platform = "*nix"}

[tool.poetry.dev-dependencies]
black = "^18.3-alpha.0"
isort = "^4.3"
pylava = "^0.2.2"
